requirements.txt
requests
requests-ip-rotator
python-dotenv
